KVM: MIPS: Sign extend KVM_GET_REGS fields
The MIPS struct kvm_vcpu_arch stores GPRs, HI, LO, and PC as unsigned
longs. Sadly this means KVM_GET_REGS on 32-bit kernels zero extend the
values into the 64-bit fields in struct kvm_regs, resulting in
non-canonical 32-bit values which aren't portable to MIPS64 state.

It was always the intention for these values to be sign extended, for
example see the comment above struct kvm_regs which was added in
688cded320a4 ("mips/kvm: Fix ABI for use of 64-bit registers."),
therefore cast the values to long so they sign extend to __u64.

This is only really likely to impact the use of MIPS64 targetting
programs running on 32-bit kernels, for example qemu-system-mips64el
reports an initial PC of 0x00000000bfc00000 instead of
0xffffffffbfc00000. Since the kernel internal state is still 32-bits the
zero extension would get truncated when the registers are written back
to KVM, unless the state is migrated to a 64-bit host kernel of course.
Any MIPS64 interpretation of the state from userland could also be
broken.

Fixes: 688cded320a4 ("mips/kvm: Fix ABI for use of 64-bit registers.")
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: "Radim Krčmář" <rkrcmar@redhat.com>
Cc: David Daney <david.daney@cavium.com>
Cc: Ralf Baechle <ralf@linux-mips.org>
Cc: linux-mips@linux-mips.org
Cc: kvm@vger.kernel.org
Cc: <stable@vger.kernel.org> # 3.10+

+++ arch/mips/kvm/mips.c
@@ -1160,11 +1160,11 @@ int kvm_arch_vcpu_ioctl_get_regs(struct kvm_vcpu *vcpu, struct kvm_regs *regs)
 	int i;
 
 	for (i = 0; i < ARRAY_SIZE(vcpu->arch.gprs); i++)
-		regs->gpr[i] = vcpu->arch.gprs[i];
+		regs->gpr[i] = (long)vcpu->arch.gprs[i];
 
-	regs->hi = vcpu->arch.hi;
-	regs->lo = vcpu->arch.lo;
-	regs->pc = vcpu->arch.pc;
+	regs->hi = (long)vcpu->arch.hi;
+	regs->lo = (long)vcpu->arch.lo;
+	regs->pc = (long)vcpu->arch.pc;
 
 	return 0;
 }
